refactor(util): extract getHost and getEnv utils
