+++ src/util/index.js
@@ -6,9 +6,19 @@ const {spawn} = require('child_process');
 const tar = require('tar-fs');
 const rimraf = require('rimraf');
 const uuid = require('uuid');
+const {getSecretsCollection} = require('../db/secrets');
 
 // our modules
 const {tempDockerDir: tempDir} = require('../config');
+
+// try to find secret with current value name and return secret value if present
+const valueOrSecret = (value, secrets) => {
+  const secret = secrets.find(s => `@${s.name}` === value);
+  if (secret) {
+    return secret.value;
+  }
+  return value;
+};
 
 // cleanup temp folder
 exports.cleanTemp = folder => new Promise(resolve => rimraf(path.join(tempDir, folder), resolve));
@@ -80,4 +90,30 @@ exports.compareNames = (nameOne = '', nameTwo = '') => {
   const nameTwoParts = nameTwo.split('-');
   const nameTwoClean = nameTwoParts.slice(0, nameTwoParts.length - 2).join('-');
   return nameOneClean === nameTwoClean;
+};
+
+exports.getHost = ({serverConfig, name, config}) => {
+  // construct base domain from config, prepend with "." if it's not there
+  const baseDomain = serverConfig.baseDomain ? serverConfig.baseDomain.replace(/^(\.?)/, '.') : undefined;
+  // construct default domain using given base domain
+  const defaultDomain = baseDomain ? `${name}${baseDomain}` : undefined;
+  // construct host
+  const host = config.domain === undefined ? defaultDomain : config.domain;
+  return host;
+};
+
+exports.getEnv = ({username, config, name, host, project = config.project || name}) => {
+  // replace env vars values with secrets if needed
+  const secrets = getSecretsCollection().find({ user: username });
+  // generate env vars (with secrets)
+  const userEnv = config.env
+    ? Object.entries(config.env).map(([key, value]) => [key, valueOrSecret(value, secrets)])
+    : [];
+  return [
+    ...userEnv,
+    ['EXOFRAME_DEPLOYMENT', name],
+    ['EXOFRAME_USER', username],
+    ['EXOFRAME_PROJECT', project],
+    ['EXOFRAME_HOST', host],
+  ];
 };
